fixes, removed the hal uart as merged into chibios
