+++ targets/ChibiOS/RP_PICO_RP2040/CMakeLists.txt
@@ -9,6 +9,10 @@ FetchContent_GetProperties(chibios)
 include(binutils.common)
 include(binutils.ChibiOS)
 
+if(NOT DEFINED RP2040_COMMON OR RP2040_COMMON STREQUAL "")
+    set(RP2040_COMMON "${CMAKE_SOURCE_DIR}/targets/ChibiOS/_RP2040")
+endif()
+
 # RP2040 target-local sources and include directories
 list(APPEND CMAKE_PREFIX_PATH "${CMAKE_CURRENT_SOURCE_DIR}")
 list(APPEND CMAKE_PREFIX_PATH "${CMAKE_CURRENT_SOURCE_DIR}/common")
@@ -18,7 +22,6 @@ list(APPEND TARGET_CHIBIOS_COMMON_INCLUDE_DIRS "${CMAKE_CURRENT_SOURCE_DIR}/comm
 
 list(APPEND TARGET_CHIBIOS_COMMON_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/Target_BlockStorage_RP2040FlashDriver.c")
 list(APPEND TARGET_CHIBIOS_COMMON_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/target_stubs.cpp")
-list(APPEND TARGET_CHIBIOS_COMMON_SOURCES "${RP2040_COMMON}/hal_uart_lld.c")
 
 nf_setup_target_build(
     HAS_NANOBOOTER
